Added method to set ids from within the page
Added a method to the embed javascript to set the blog ID and post ID from
within the body of a webpage, and have that used to be the ids that are loaded
by the javascript.  This would allow someone to set the IDs in a standalone
page hosted elsewhere on the web and have it behave as a blog-app blog that is
pseudo-self-hosted, and could be wrapped in a custom webpage and theme.